checked occurences of listFiles for NPE

+++ src/org/tinymediamanager/core/movie/MovieRenamer.java
@@ -695,18 +695,20 @@ public class MovieRenamer {
     // clean all non tmm nfos
     if (MovieModuleManager.MOVIE_SETTINGS.isMovieRenamerNfoCleanup()) {
       File[] content = new File(movie.getPath()).listFiles();
-      for (File file : content) {
-        if (file.isFile() && file.getName().toLowerCase().endsWith(".nfo")) {
-          // check if it's a tmm nfo
-          boolean supported = false;
-          for (MediaFile nfo : movie.getMediaFiles(MediaFileType.NFO)) {
-            if (nfo.getFilename().equals(file.getName())) {
-              supported = true;
+      if (content != null) {
+        for (File file : content) {
+          if (file.isFile() && file.getName().toLowerCase().endsWith(".nfo")) {
+            // check if it's a tmm nfo
+            boolean supported = false;
+            for (MediaFile nfo : movie.getMediaFiles(MediaFileType.NFO)) {
+              if (nfo.getFilename().equals(file.getName())) {
+                supported = true;
+              }
             }
-          }
-          if (!supported) {
-            LOGGER.debug("Deleting " + file);
-            FileUtils.deleteQuietly(file);
+            if (!supported) {
+              LOGGER.debug("Deleting " + file);
+              FileUtils.deleteQuietly(file);
+            }
           }
         }
       }
